+++ LongevityWorldCup.Website/Business/ImageTextLayout.cs
@@ -1,5 +1,7 @@
 using SixLabors.Fonts;
+using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.Drawing.Processing;
+using SixLabors.ImageSharp.Processing;
 
 namespace LongevityWorldCup.Website.Business;
 
@@ -17,5 +19,69 @@ internal static class ImageTextLayout
         }
 
         return family.CreateFont(minSize, FontStyle.Bold);
+    }
+
+    internal static void DrawWrappedText(
+        IImageProcessingContext ctx,
+        string text,
+        Font font,
+        Color color,
+        PointF origin,
+        float maxWidth,
+        int maxLines,
+        float lineHeight)
+    {
+        var lines = WrapText(text, font, maxWidth, maxLines);
+        for (var i = 0; i < lines.Count; i++)
+        {
+            ctx.DrawText(new RichTextOptions(font)
+            {
+                Origin = new PointF(origin.X, origin.Y + (lineHeight * i)),
+                HorizontalAlignment = HorizontalAlignment.Left,
+                VerticalAlignment = VerticalAlignment.Top
+            }, lines[i], color);
+        }
+    }
+
+    private static IReadOnlyList<string> WrapText(string text, Font font, float maxWidth, int maxLines)
+    {
+        var words = text.Split(' ', StringSplitOptions.RemoveEmptyEntries);
+        var lines = new List<string>();
+        var current = "";
+        var truncated = false;
+        for (var index = 0; index < words.Length; index++)
+        {
+            var word = words[index];
+            var candidate = string.IsNullOrWhiteSpace(current) ? word : $"{current} {word}";
+            if (TextMeasurer.MeasureSize(candidate, new RichTextOptions(font)).Width <= maxWidth)
+            {
+                current = candidate;
+                continue;
+            }
+
+            if (!string.IsNullOrWhiteSpace(current))
+                lines.Add(current);
+            current = word;
+            if (lines.Count >= maxLines)
+            {
+                truncated = index < words.Length - 1;
+                break;
+            }
+        }
+
+        if (!string.IsNullOrWhiteSpace(current) && lines.Count < maxLines)
+            lines.Add(current);
+
+        if (truncated && lines.Count == maxLines && words.Length > 0)
+        {
+            var last = lines[^1];
+            while (last.Length > 0 && TextMeasurer.MeasureSize(last + "...", new RichTextOptions(font)).Width > maxWidth)
+            {
+                last = last[..^1].TrimEnd();
+            }
+            lines[^1] = last + "...";
+        }
+
+        return lines;
     }
 }
